+++ src/orders/components/OrderCardTitle/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./OrderCardTitle";
+export * from "./OrderCardTitle";
